pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "codoo"
version = "1.0.0"
description = "Deterministic, AI-assisted Odoo development with full traceability"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "Open2 Technology", email = "hello@open2.tech" }
]
keywords = ["odoo", "automation", "ai-agents", "deterministic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
]

dependencies = [
    "pydantic>=2.10.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "typer[all]>=0.14.0",
    "pyyaml>=6.0.1",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.14.0",
    "black>=24.10.0",
    "ruff>=0.8.0",
    "pytest-mock>=3.14.0",
]

[project.urls]
Repository = "https://github.com/marcelo-m7/facodi"
Documentation = "https://open2.tech"
Issues = "https://github.com/marcelo-m7/facodi/issues"

[project.scripts]
codoo = "codoo.__main__:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/codoo"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
addopts = "--cov=src/codoo --cov-report=term-missing --cov-report=html"

[tool.coverage.run]
source = ["src/codoo"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
exclude = ["tests/", ".venv"]

[tool.black]
line-length = 100
target-version = ["py313"]
include = '\.pyi?$'

[tool.ruff]
target-version = "py313"
line-length = 100
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "C",    # flake8-comprehensions
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.isort]
known-first-party = ["codoo"]
